Add a controlld state on form
